Move store to other file
