+++ resources/js/store.js
@@ -1,0 +1,9 @@
+import { createStore, applyMiddleware } from "redux";
+import thunkMiddleware from "redux-thunk";
+import { composeWithDevTools } from "redux-devtools-extension";
+import { rootReducer } from "./reducers";
+
+export const store = createStore(
+    rootReducer,
+    composeWithDevTools(applyMiddleware(thunkMiddleware))
+);
